+++ cmd/notes/new.go
@@ -1,12 +1,12 @@
 package main
 
 import (
+	"fmt"
 	"time"
 
 	"github.com/subtlepseudonym/notes"
 	dalpkg "github.com/subtlepseudonym/notes/dal"
 
-	"github.com/pkg/errors"
 	"github.com/urfave/cli"
 	"go.uber.org/zap"
 )
@@ -28,6 +28,10 @@ func buildNewCommand(dal dalpkg.DAL, meta *notes.Meta) cli.Command {
 			cli.BoolFlag{
 				Name:  "no-watch",
 				Usage: "don't save note in background",
+			},
+			cli.BoolFlag{
+				Name:  "no-history",
+				Usage: "don't record activity in edit history",
 			},
 			cli.StringFlag{
 				Name:  "title, t",
@@ -59,17 +63,19 @@ func buildNewCommand(dal dalpkg.DAL, meta *notes.Meta) cli.Command {
 }
 
 func newAction(ctx *cli.Context, dal dalpkg.DAL, meta *notes.Meta) error {
+	logger := zap.L().Named(ctx.Command.Name)
+
 	newNoteID := meta.LatestID + 1
 	_, exists := meta.Notes[newNoteID]
 	if exists {
-		return cli.NewExitError(errors.New("note ID is not unique"), 1)
+		return fmt.Errorf("note ID: must be unique")
 	}
 
 	var title string
 	if ctx.String("title") != "" {
 		title = ctx.String("title")
 	} else {
-		title = generateDateTitle(ctx.String("title-format"), ctx.String("title-location"))
+		title = generateDateTitle(ctx.String("title-format"), ctx.String("title-location"), logger)
 	}
 
 	note := &notes.Note{
@@ -84,42 +90,45 @@ func newAction(ctx *cli.Context, dal dalpkg.DAL, meta *notes.Meta) error {
 
 	body, err := editNote(ctx, dal, meta, note)
 	if err != nil {
-		return cli.NewExitError(errors.Wrap(err, "user handoff failed"), 1)
+		return fmt.Errorf("user handoff: %w", err)
 	}
 	note.Body = body
 
-	// TODO: add option to not append edit to history
-	note, err = note.AppendEdit(time.Now())
-	if err != nil {
-		return cli.NewExitError(errors.Wrap(err, "append edit to note history failed"), 1)
+	if !ctx.Bool("no-history") {
+		note, err = note.AppendEdit(time.Now())
+		if err != nil {
+			return fmt.Errorf("append edit to note history: %w", err)
+		}
 	}
 
 	err = dal.SaveNote(note)
 	if err != nil {
 		// FIXME: persist the note somewhere if saving it fails
-		return cli.NewExitError(errors.Wrap(err, "save note failed"), 1)
+		return fmt.Errorf("save note: %w", err)
 	}
+	logger.Info("note updated", zap.Int("noteID", note.Meta.ID))
 
 	metaSize, err := meta.ApproxSize()
 	if err != nil {
-		return cli.NewExitError(errors.Wrap(err, "get meta size failed"), 1)
+		return fmt.Errorf("get meta size: %w", err)
 	}
 
 	meta.Size = metaSize
 	meta.Notes[note.Meta.ID] = note.Meta
 	err = dal.SaveMeta(meta)
 	if err != nil {
-		return cli.NewExitError(errors.Wrap(err, "save meta failed"), 1)
+		return fmt.Errorf("save meta: %w", err)
 	}
+	logger.Info("meta updated", zap.Int("metaSize", meta.Size))
 
 	return nil
 }
 
-func generateDateTitle(format, location string) string {
+func generateDateTitle(format, location string, logger *zap.Logger) string {
 	var loc *time.Location
 	l, err := time.LoadLocation(location)
 	if err != nil {
-		Logger.Warn("load location failed, defaulting to UTC", zap.String("location", location), zap.String("format", format), zap.Error(err))
+		logger.Warn("load location failed, defaulting to UTC", zap.String("location", location), zap.String("format", format), zap.Error(err))
 		loc = time.UTC
 	} else {
 		loc = l