+++ cmd/notes/rm.go
@@ -1,14 +1,15 @@
 package main
 
 import (
+	"fmt"
 	"strconv"
 	"time"
 
 	"github.com/subtlepseudonym/notes"
 	dalpkg "github.com/subtlepseudonym/notes/dal"
 
-	"github.com/pkg/errors"
 	"github.com/urfave/cli"
+	"go.uber.org/zap"
 )
 
 func buildRemoveCommand(dal dalpkg.DAL, meta *notes.Meta) cli.Command {
@@ -29,24 +30,26 @@ func buildRemoveCommand(dal dalpkg.DAL, meta *notes.Meta) cli.Command {
 }
 
 func rmAction(ctx *cli.Context, dal dalpkg.DAL, meta *notes.Meta) error {
+	logger := zap.L().Named(ctx.Command.Name)
+
 	if !ctx.Args().Present() {
-		return cli.NewExitError(errors.New("note ID argument is required"), 1)
+		return fmt.Errorf("usage: noteID argument required")
 	}
 	n, err := strconv.ParseInt(ctx.Args().First(), 16, 64)
 	if err != nil {
-		return cli.NewExitError(errors.Wrap(err, "parse base 16 noteID argument failed"), 1)
+		return fmt.Errorf("parse noteID argument: %w", err)
 	}
 	noteID := int(n)
 
 	note, err := dal.GetNote(noteID)
 	if err != nil {
-		return cli.NewExitError(errors.Wrap(err, "get note failed"), 1)
+		return fmt.Errorf("get note: %w", err)
 	}
 
 	if ctx.Bool("hard") {
 		err = dal.RemoveNote(noteID)
 		if err != nil {
-			return cli.NewExitError(errors.Wrap(err, "remove note file failed"), 1)
+			return fmt.Errorf("remove note file: %w", err)
 		}
 
 		delete(meta.Notes, note.Meta.ID)
@@ -54,15 +57,16 @@ func rmAction(ctx *cli.Context, dal dalpkg.DAL, meta *notes.Meta) error {
 		note.Meta.Deleted.Time = time.Now()
 		err = dal.SaveNote(note)
 		if err != nil {
-			return cli.NewExitError(errors.Wrap(err, "save note failed"), 1)
+			return fmt.Errorf("save note: %w", err)
 		}
+		logger.Info("note updated", zap.Int("noteID", note.Meta.ID))
 
 		meta.Notes[note.Meta.ID] = note.Meta
 	}
 
 	err = dal.SaveMeta(meta)
 	if err != nil {
-		return cli.NewExitError(errors.Wrap(err, "save meta failed"), 1)
+		return fmt.Errorf("save meta: %w", err)
 	}
 
 	return nil