+++ dal/dal.go
@@ -9,7 +9,6 @@ import (
 	"github.com/subtlepseudonym/notes"
 
 	"github.com/mitchellh/go-homedir"
-	"github.com/pkg/errors"
 )
 
 const (
@@ -39,7 +38,7 @@ type localDAL struct {
 func NewLocalDAL(dirName, version string) (DAL, error) {
 	home, err := homedir.Dir()
 	if err != nil {
-		return nil, errors.Wrap(err, "get home directory failed")
+		return nil, fmt.Errorf("get home directory: %v", err)
 	}
 
 	return &localDAL{
@@ -54,14 +53,14 @@ func (d *localDAL) buildNewMeta() (*notes.Meta, error) {
 	if _, err := os.Stat(d.notesDirectoryPath); os.IsNotExist(err) {
 		err = os.Mkdir(d.notesDirectoryPath, os.ModeDir|os.FileMode(0700))
 		if err != nil {
-			return nil, errors.Wrap(err, "create notes directory failed")
+			return nil, fmt.Errorf("create notes directory: %v", err)
 		}
 	}
 
 	metaPath := path.Join(d.notesDirectoryPath, d.metaFilename)
 	metaFile, err := os.Create(metaPath)
 	if err != nil {
-		return nil, errors.Wrap(err, "create meta file failed")
+		return nil, fmt.Errorf("create meta file: %v", err)
 	}
 	defer metaFile.Close()
 
@@ -72,10 +71,14 @@ func (d *localDAL) buildNewMeta() (*notes.Meta, error) {
 
 	err = json.NewEncoder(metaFile).Encode(m)
 	if err != nil {
-		return nil, errors.Wrap(err, "encode meta file failed")
+		return nil, fmt.Errorf("encode meta file: %v", err)
 	}
 
-	return m, errors.Wrap(metaFile.Close(), "close meta file failed")
+	err = metaFile.Close()
+	if err != nil {
+		return m, fmt.Errorf("close meta file: %v", err)
+	}
+	return m, nil
 }
 
 // GetMeta retrieves and decodes a Meta from file
@@ -90,10 +93,14 @@ func (d *localDAL) GetMeta() (*notes.Meta, error) {
 	var m notes.Meta
 	err = json.NewDecoder(metaFile).Decode(&m)
 	if err != nil {
-		return nil, errors.Wrap(err, "decode meta file failed")
+		return nil, fmt.Errorf("decode meta file: %v", err)
 	}
 
-	return &m, errors.Wrap(metaFile.Close(), "close meta file failed")
+	err = metaFile.Close()
+	if err != nil {
+		return &m, fmt.Errorf("close meta file: %v", err)
+	}
+	return &m, nil
 }
 
 // SaveMeta encodes and saves the provided Meta to file
@@ -101,7 +108,7 @@ func (d *localDAL) SaveMeta(meta *notes.Meta) error {
 	metaPath := path.Join(d.notesDirectoryPath, d.metaFilename)
 	err := os.Rename(metaPath, metaPath+".bak")
 	if err != nil {
-		return errors.Wrap(err, "backup old meta file failed")
+		return fmt.Errorf("backup old meta file: %v", err)
 	}
 	// TODO: remove meta backup
 
@@ -109,18 +116,22 @@ func (d *localDAL) SaveMeta(meta *notes.Meta) error {
 	if err != nil {
 		err = os.Rename(metaPath+".bak", metaPath)
 		if err != nil {
-			return errors.Wrap(err, "restoring meta backup failed")
+			return fmt.Errorf("restore meta backup: %v", err)
 		}
-		return errors.Wrap(err, "create meta file failed")
+		return fmt.Errorf("create meta file: %v", err)
 	}
 	defer metaFile.Close()
 
 	err = json.NewEncoder(metaFile).Encode(meta)
 	if err != nil {
-		return errors.Wrap(err, "encode meta file failed")
+		return fmt.Errorf("encode meta file: %v", err)
 	}
 
-	return errors.Wrap(metaFile.Close(), "close meta file failed")
+	err = metaFile.Close()
+	if err != nil {
+		return fmt.Errorf("close meta file: %v", err)
+	}
+	return nil
 }
 
 func (d *localDAL) getNotePath(id int) string {
@@ -133,17 +144,21 @@ func (d *localDAL) GetNote(id int) (*notes.Note, error) {
 	notePath := d.getNotePath(id)
 	noteFile, err := os.Open(notePath)
 	if err != nil {
-		return nil, errors.Wrap(err, "open note file failed")
+		return nil, fmt.Errorf("open note file: %v", err)
 	}
 	defer noteFile.Close()
 
 	var n notes.Note
 	err = json.NewDecoder(noteFile).Decode(&n)
 	if err != nil {
-		return nil, errors.Wrap(err, "decode note file failed")
+		return nil, fmt.Errorf("decode note file: %v", err)
 	}
 
-	return &n, errors.Wrap(noteFile.Close(), "close note file failed")
+	err = noteFile.Close()
+	if err != nil {
+		return &n, fmt.Errorf("close note file: %v", err)
+	}
+	return &n, nil
 }
 
 // SaveNote encodes and saves the provided Note to file
@@ -151,17 +166,21 @@ func (d *localDAL) SaveNote(note *notes.Note) error {
 	notePath := d.getNotePath(note.Meta.ID)
 	noteFile, err := os.Create(notePath)
 	if err != nil {
-		return errors.Wrap(err, "create note file failed")
+		return fmt.Errorf("create note file: %v", err)
 	}
 	defer noteFile.Close()
 
 	err = json.NewEncoder(noteFile).Encode(note)
 	if err != nil {
 		os.Remove(notePath) // FIXME: do something with this error
-		return errors.Wrap(err, "encode note file failed")
+		return fmt.Errorf("encode note file: %v", err)
 	}
 
-	return errors.Wrap(noteFile.Close(), "close note file failed")
+	err = noteFile.Close()
+	if err != nil {
+		return fmt.Errorf("close note file: %v", err)
+	}
+	return nil
 }
 
 // RemoveNote deletes the note file
@@ -169,7 +188,7 @@ func (d *localDAL) RemoveNote(id int) error {
 	notePath := d.getNotePath(id)
 	err := os.Remove(notePath)
 	if err != nil {
-		return errors.Wrap(err, "remove note file failed")
+		return fmt.Errorf("remove note file: %v", err)
 	}
 
 	return nil