+++ notes.go
@@ -3,22 +3,27 @@ package notes
 import (
 	"encoding/json"
 	"fmt"
-	"io/ioutil"
-	"os"
-	"os/exec"
 	"time"
-
-	"github.com/pkg/errors"
 )
 
 const defaultEditHistorySize = 16
 
 // Meta holds meta information for the local notes storage as a whole
 type Meta struct {
-	Version  string           `json:"version"`
-	LatestID int              `json:"latestId"`
-	Size     int              `json:"size"`  // meta file size in bytes
-	Notes    map[int]NoteMeta `json:"notes"` // maps note ID to NoteMeta
+	Version     string           `json:"version"`
+	OldVersions []string         `json:"oldVersions"`
+	LatestID    int              `json:"latestId"`
+	Size        int              `json:"size"`  // meta file size in bytes
+	Notes       map[int]NoteMeta `json:"notes"` // maps note ID to NoteMeta
+}
+
+// UpdateVersion replaces the existing version with the provided new version
+// and adds the old version to the list of old versions
+func (m *Meta) UpdateVersion(newVersion string) *Meta {
+	m.OldVersions = append(m.OldVersions, m.Version)
+	m.Version = newVersion
+
+	return m
 }
 
 // ApproxSize gets the approximate encoded size of the meta object by
@@ -28,7 +33,7 @@ type Meta struct {
 func (m Meta) ApproxSize() (int, error) {
 	b, err := json.Marshal(m)
 	if err != nil {
-		return 0, errors.Wrap(err, "encode meta failed")
+		return 0, fmt.Errorf("encode meta: %w", err)
 	}
 
 	return len(b), nil
@@ -84,7 +89,7 @@ type Note struct {
 func (n Note) ApproxSize() (int, error) {
 	b, err := json.Marshal(n)
 	if err != nil {
-		return 0, errors.Wrap(err, "encode note failed")
+		return 0, fmt.Errorf("encode note: %w", err)
 	}
 
 	return len(b), nil
@@ -95,7 +100,7 @@ func (n Note) ApproxSize() (int, error) {
 func (n *Note) AppendEdit(timestamp time.Time) (*Note, error) {
 	noteSize, err := n.ApproxSize()
 	if err != nil {
-		return n, errors.Wrap(err, "get note size failed")
+		return n, fmt.Errorf("get note size: %w", err)
 	}
 
 	update := EditHistory{
@@ -108,29 +113,4 @@ func (n *Note) AppendEdit(timestamp time.Time) (*Note, error) {
 	}
 
 	return n, nil
-}
-
-// GetNoteBodyFromUser drops the user into the provided editor command before
-// retrieving the contents of the edited file
-func GetNoteBodyFromUser(file *os.File, editor, existingBody string) (string, error) {
-	_, err := fmt.Fprint(file, existingBody)
-	if err != nil {
-		return "", errors.Wrap(err, "print existing body to temporary file failed")
-	}
-
-	cmd := exec.Command(editor, file.Name())
-	cmd.Stdin = os.Stdin
-	cmd.Stdout = os.Stdout
-
-	err = cmd.Run()
-	if err != nil {
-		return "", errors.Wrap(err, "run editor command failed")
-	}
-
-	bodyBytes, err := ioutil.ReadFile(file.Name())
-	if err != nil {
-		return "", errors.Wrap(err, "read temporary file failed")
-	}
-
-	return string(bodyBytes), nil
 }